kream-crawling/requirements.txt
selenium>=4.25.0
pandas>=2.2.2
python-dotenv>=1.0.1